refactor: delete save answer command class AnsweredQuestion
